# Improve performance and reduce allocations in DecisionTreeBuilder<T>.