Improves performance and reduces allocations in DecisionTreeBuilder<T>.

**Before:**

|       Method |     Mean |     Error |    StdDev |     Op/s |  Gen 0 | Allocated |
|------------- |---------:|----------:|----------:|---------:|-------:|----------:|
| GenerateTree | 18.56 us | 0.0332 us | 0.0310 us | 53,886.4 | 0.2747 |  23.86 KB |

**After**:

|       Method |     Mean |     Error |    StdDev |     Op/s |  Gen 0 | Allocated |
|------------- |---------:|----------:|----------:|---------:|-------:|----------:|
| GenerateTree | 16.91 us | 0.0689 us | 0.0611 us | 59,119.2 | 0.2136 |  21.02 KB |

Benchmark is available [here](https://gist.github.com/drieseng/2bfa1345c22850c180b4b3ea6b18bbc3).